PEP 639 license metadata
License classifiers are deprecated.

`license` field is now plain text (SPDX identifier).

`license-files` field added, supposed to reference the license file(s)
included with the code ("COPYING" in this case).

Only `setuptools` 77+ supports this standard. It does not work *to
build* on Python 3.8, but the resulting package should be installable.
(Minimum is set to 77.0.3 because of bugs in the earlier versions.)