Update `atproto` and enforce newer `pydantic`, where compatible
A lot of projects have dropped Python 3.8 at this point, but if possible
it's ideal for this plugin to have newer versions of `pydantic` in
particular due to performance improvements in v2.11.x.

Python 3.8 can't benefit from that, but let's at least make sure it gets
the newest compatible `pydantic` (2.10.x) and the last version of
`atproto` that's supported.