Don't resolve handle before fetching a skeet
Bluesky doesn't force canonical URLs, and thus it's possible to get post
URLs containing either the user's DID or handle depending on... factors.

Trying to `client.resolve_handle()` a string that's already a DID breaks
pretty spectacularly, but the good news is that `client.get_post()` and
`client.get_profile()` work just fine with either kind of input when
specifying the user.

This could have been a ±1 patch except for not wanting to leave the
`handle` capture group and `did` local variable with those names that
don't correctly reflect what they hold.